feat(tax-advisor): update metadata and hero section content
Refactor the tax advisor section to simplify the messaging and 
improve clarity. Update titles, descriptions, and call-to-action 
to promote personalized tax advice and enhance user engagement. 
Revise statistics presentation and change the button link for 
better navigation.
