--- app/dictionaries.ts
@@ -1,9 +1,0 @@
-import 'server-only'
- 
-const dictionaries = {
-  en: () => import('@/dictionaries/en.json').then((module) => module.default),
-  nl: () => import('@/dictionaries/nl.json').then((module) => module.default),
-}
- 
-export const getDictionary = async (locale: 'en' | 'nl') =>
-  dictionaries[locale]()
