+++ app/en/privacy/page.tsx
@@ -33,7 +33,7 @@ export default async function LocalizedPrivacyPage({
   const dict = await getDictionary("en");
   return (
     <main className="flex-1 flex flex-col relative">
-      <Header />
+      <Header dict={dict} />
       <article className="max-w-4xl prose text-gray-500 mx-auto">
         <h2>Privacy Policy</h2>
         <p>
