+++ app/onderwerpen/[topic]/page.tsx
@@ -1,3 +1,4 @@
+import { getDictionary } from "@/app/dictionaries";
 import Footer from "@/components/Footer";
 import Header from "@/components/Header";
 import FAQAndBlogSection from "@/components/sections/FAQAndBlogSection";
@@ -5,7 +6,6 @@ import HeroSection from "@/components/sections/HeroSection";
 import PartnersSection from "@/components/sections/PartnersSection";
 import StepsSection from "@/components/sections/StepsSection";
 import SupportSection from "@/components/sections/SupportSection";
-import { getDictionary } from "@/lib/getDictionary";
 import { generatePageMetadata } from "@/lib/metadata";
 import { topicENToDictionaryKey, topicNLToDictionaryKey } from "@/lib/routes";
 import { SupportedLocale } from "@/lib/types";
@@ -15,80 +15,6 @@ interface TopicPageProps {
   params: {
     topic: string;
     locale?: string;
-  };
-}
-
-type Dictionary = Awaited<ReturnType<typeof getDictionary>>;
-type DictionaryKeys = keyof Dictionary["topics"];
-
-interface Partner {
-  src: string;
-  alt: string;
-  width: number;
-}
-
-interface Step {
-  title: string;
-  description: string;
-  image: string;
-  alt: string;
-  showButton?: boolean;
-}
-
-interface PageContent {
-  metadata: {
-    title: string;
-    description: string;
-    ogTitle: string;
-    ogDescription: string;
-    ogImageAlt: string;
-  };
-  hero: {
-    title: string;
-    description: string;
-    callToAction: string;
-    buttonText: string;
-    buttonSubtext: string;
-    buttonLink: string;
-    image: {
-      src: string;
-      alt: string;
-      width: number;
-      height: number;
-    };
-    imageAlt: string;
-    stats: string;
-  };
-  partners: {
-    title: string;
-    items: Partner[];
-  };
-  support: {
-    title: string;
-    description: string;
-    buttonText: string;
-    buttonSubtext: string;
-    buttonLink: string;
-    imageSrc: string;
-    imageAlt: string;
-  };
-  stepsSection: {
-    buttonText: string;
-    buttonLink: string;
-    buttonSubtext: string;
-    items: Step[];
-  };
-  faq: {
-    title: string;
-    faqTitle: string;
-    blogTitle: string;
-    faqItems: Array<{
-      question: string;
-      answer: string;
-    }>;
-    blogCategory: string;
-    buttonText: string;
-    buttonLink: string;
   };
 }
 
@@ -117,15 +43,15 @@ export async function generateStaticParams() {
 export const generateMetadata = generatePageMetadata;
 
 export default async function TopicPage({ params }: TopicPageProps) {
-  const { topic, locale = "nl" } = await params;
+  const { topic, locale = "nl" } = params;
   const dict = await getDictionary(locale as SupportedLocale);
   let dictionaryKey: string;
+
   // Get the dictionary key for this topic
-  console.log("topic", topic);
   if (locale === "nl") {
-    dictionaryKey = topicNLToDictionaryKey[topic];
+    dictionaryKey = topicNLToDictionaryKey[topic] as string;
   } else {
-    dictionaryKey = topicENToDictionaryKey[topic];
+    dictionaryKey = topicENToDictionaryKey[topic] as string;
   }
 
   // If we don't have a mapping for this topic, show 404
@@ -135,9 +61,7 @@ export default async function TopicPage({ params }: TopicPageProps) {
   }
 
   // Get the topic-specific content using the dictionary key
-  const topicContent = dict.topics[
-    dictionaryKey as DictionaryKeys
-  ] as PageContent;
+  const topicContent = dict.topics[dictionaryKey];
 
   // If the topic content doesn't exist, show 404
   if (!topicContent) {
@@ -145,47 +69,45 @@ export default async function TopicPage({ params }: TopicPageProps) {
     notFound();
   }
 
-  const { hero, partners, support, faq } = topicContent;
+  const { hero, partners, support, stepsSection, faq } = topicContent;
 
-  console.log("hero", hero);
-
+  console.log("topicContent", topicContent);
   return (
-    <main className="flex-1 flex flex-col relative">
-      <Header />
+    <main className="relative flex-1 flex flex-col pt-20">
+      <Header dict={dict} />
       <div className="pt-22 overflow-y-auto h-[calc(100vh)]">
         <div className="max-w-5xl mx-auto px-6 py-12">
-          <HeroSection
-            title={hero.title}
-            description={hero.description}
-            callToAction={hero.callToAction}
-            buttonText={hero.buttonText}
-            buttonSubtext={hero.buttonSubtext}
-            buttonLink={hero.buttonLink}
-            image={hero.image}
-            stats={hero.stats}
-          />
+          <HeroSection {...hero} />
           <PartnersSection title={partners.title} partners={partners.items} />
-          <SupportSection {...support} />
+          <SupportSection
+            title={support.title}
+            description={support.description}
+            buttonText={support.buttonText}
+            buttonSubtext={support.buttonSubtext}
+            buttonLink={support.buttonLink}
+            imageSrc={support.imageSrc}
+            imageAlt={support.imageAlt}
+          />
           <StepsSection
-            steps={topicContent.stepsSection.items}
-            buttonText={topicContent.stepsSection.buttonText}
-            buttonLink={topicContent.stepsSection.buttonLink}
-            buttonSubtext={topicContent.stepsSection.buttonSubtext}
+            steps={stepsSection.items}
+            buttonText={stepsSection.buttonText}
+            buttonLink={stepsSection.buttonLink}
+            buttonSubtext={stepsSection.buttonSubtext}
           />
-          <FAQAndBlogSection
-            title={faq.title}
-            faqTitle={faq.faqTitle}
-            blogTitle={faq.blogTitle}
-            blogCategory={faq.blogCategory}
-            faqItems={faq.faqItems}
-            buttonText={faq.buttonText}
-            buttonLink={faq.buttonLink}
-          />
+          <div className="container">
+            <FAQAndBlogSection
+              title={faq.title}
+              faqTitle={faq.faqTitle}
+              blogTitle={faq.blogTitle}
+              faqItems={faq.faqItems}
+              blogCategory={faq.blogCategory || ""}
+              buttonText={faq.buttonText}
+              buttonLink={faq.buttonLink}
+            />
+          </div>
         </div>
       </div>
       <Footer dict={dict} />
     </main>
   );
 }
-
-
