+++ app/over-ons/page.tsx
@@ -78,7 +78,7 @@ export default async function OverOnsPage({ params }: OverOnsPageProps) {
 
   return (
     <main className="relative flex-1 flex flex-col pt-20">
-      <Header />
+      <Header dict={dict} />
       <div className="flex-1 overflow-y-auto">
         <div className="max-w-5xl mx-auto px-6 py-12">
           <AboutUsContent dict={aboutUs} />
